Remove Pi::planner and make a class for it
Decouple TransferPlanner from other classes
and let it be managed more directly
