+++ src/Game.h
@@ -16,11 +16,11 @@
 class Galaxy;
 class GameLog;
 class HyperspaceCloud;
+class InGameViews;
 class LuaTimer;
 class Player;
 class Space;
-class InGameViews;
-class ShipCpanel;
+class TransferPlanner;
 
 namespace Graphics {
 	class Renderer;
@@ -112,6 +112,7 @@ public:
 
 	static void EmitPauseState(bool paused);
 
+	const TransferPlanner *GetPlanner() const;
 	InGameViews *GetInGameViews() { return m_inGameViews.get(); };
 
 private:
@@ -132,8 +133,6 @@ private:
 	double m_time;
 
 	std::unique_ptr<Player> m_player;
-
-
 	std::unique_ptr<LuaTimer> m_luaTimer;
 
 	enum class State {
